app/requirements.txt
flask
python-decouple
snowflake-connector-python